requirements.txt
asgiref==3.6.0
Django==4.1.5
django-ipware==4.0.2
django-taggit==3.1.0
psycopg2-binary==2.9.5
python-dotenv==0.21.1
sqlparse==0.4.3
